fix multi-logging (#30)
When a test was failing in a rerun, the test appeared multiple times in the failures files. This was due to the fact that multiple instances of the same listener were added to surefire, which resulted in multiple loggings of the same failing test.

The fix was to add the listener only in the first surefire invocation.